feat: Update configuration parameters and image transformation
The configuration parameters in `configuration.py` have been updated to set the `TEST_SIZE` to 0.2 and `EPOCHS` to 15. This change improves the training process. Additionally, the `BilateralFilter` class in `image_transformation.py` has been modified to use different values for the `d`, `sigmaColor`, and `sigmaSpace` parameters, resulting in better image filtering.

Co-authored-by: EnriqueVilchezL <EnriqueVilchezL@users.noreply.github.com>